Moved downsampling functionality to a simple_downsample function
